chore(git): ignore generated local documentation files at cmd/docs/*.md
It is convenient to hide these files from the git file changes as
feature development usually requires adjusting the documentation as
well, which can be easily verified with the generated local documentation.

These documentation files are also changing dynamically which would
require extra efforts to be kept in sync on the repository remote thus
committing them would not be feasible.